Revert "refactor: Split API to smalest parts"
This reverts commit 10e497b26cb1bd772848f27b85c528a8a0f20aa9.
