+++ src/components/ProductCard/ProductCard.tsx
@@ -1,16 +1,16 @@
-import { ProductProjection } from '@commercetools/platform-sdk';
+import { Product } from '@commercetools/platform-sdk';
 import './ProductCard.css';
 import { Link } from 'react-router-dom';
 
 interface ProductCardProps {
-  product: ProductProjection;
+  product: Product;
 }
 
 const ProductCard: React.FC<ProductCardProps> = ({ product }) => {
-  const masterVariant = product.masterVariant;
+  const masterVariant = product.masterData.current.masterVariant;
   const firstImage = masterVariant.images?.[0];
-  const productName = product.name['en-US'];
-  const productDescription = product.description?.['en-US'] || '';
+  const productName = product.masterData.current.name['en-US'];
+  const productDescription = product.masterData.current.description?.['en-US'] || '';
 
   const priceData = masterVariant.prices?.[0];
   const originalPriceCents = priceData?.value?.centAmount;
