+++ src/components/products-list/Products-list.css
@@ -40,19 +40,6 @@
   padding: 40px 0;
 }
 
-.search-button {
-  padding: 8px 16px;
-  /* background-color: #007bff; */
-  /* color: white; */
-  /* border: none; */
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-/* .search-button:hover {
-  background-color: #0056b3;
-} */
-
 .loading,
 .error {
   text-align: center;
