+++ src/components/products-list/Products-list.tsx
@@ -1,24 +1,23 @@
 import './Products-list.css';
 import { useEffect, useState } from 'react';
 import api from '@/api/api';
-import { ProductProjectionPagedSearchResponse } from '@commercetools/platform-sdk';
+import { ProductPagedQueryResponse } from '@commercetools/platform-sdk';
 import ProductCard from '@/components/ProductCard/ProductCard';
 import PaginationControls from '@/components/products-list/PaginationControls';
-import { ProductProjectionSearchArgs } from '@/api/interfaces/types';
+import { ProductsQueryArgs } from '@/api/api-interfaces';
 
 const ProductList: React.FC = () => {
-  const [products, setProducts] = useState<ProductProjectionPagedSearchResponse | null>(null);
+  const [products, setProducts] = useState<ProductPagedQueryResponse | null>(null);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [searchQuery, setSearchQuery] = useState('');
-  const [appliedSearchQuery, setAppliedSearchQuery] = useState('');
   const [selectedCategory, setSelectedCategory] = useState<string>('');
   const [offset, setOffset] = useState(0);
   const [limit, setLimit] = useState<number>(5);
 
   const limitOptions = [5, 10, 15, 25, 30, 50, 100];
   const categories = [
-    { id: '1', name: 'Electronics' },
+    { id: '1', name: 'Electronics' }, // todo change to real
     { id: '2', name: 'Clothing' },
     { id: '3', name: 'Books' },
   ];
@@ -29,15 +28,16 @@ const ProductList: React.FC = () => {
         setLoading(true);
         setError(null);
 
-        const params: ProductProjectionSearchArgs = {
+        const params: ProductsQueryArgs = {
           limit,
           offset,
-          ...(appliedSearchQuery && { searchTerm: appliedSearchQuery }),
-          filter: selectedCategory ? [`categories.id:"${selectedCategory}"`] : undefined,
-          facet: ['variants.attributes.color', 'variants.attributes.size'],
+          ...(searchQuery && { search: searchQuery }),
+          ...(selectedCategory && { category: selectedCategory }),
         };
 
         const response = await api.getProductsList(params);
+        console.log('params = ', params);
+        console.log('response = ', response);
 
         if (response) {
           setProducts(response.body);
@@ -51,20 +51,10 @@ const ProductList: React.FC = () => {
     };
 
     fetchProducts();
-  }, [appliedSearchQuery, selectedCategory, offset, limit]);
+  }, [searchQuery, selectedCategory, offset, limit]);
 
   const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setSearchQuery(e.target.value);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      applySearch();
-    }
-  };
-
-  const applySearch = () => {
-    setAppliedSearchQuery(searchQuery);
     setOffset(0);
   };
 
@@ -94,18 +84,12 @@ const ProductList: React.FC = () => {
       <h2>Product List</h2>
 
       <div className="product-filters">
-        <div className="search-container">
-          <input
-            type="text"
-            placeholder="Search products..."
-            value={searchQuery}
-            onChange={handleSearchChange}
-            onKeyDown={handleKeyDown}
-          />
-          <button onClick={applySearch} className="search-button">
-            Search
-          </button>
-        </div>
+        <input
+          type="text"
+          placeholder="Search products..."
+          value={searchQuery}
+          onChange={handleSearchChange}
+        />
 
         <select value={selectedCategory} onChange={handleCategoryChange}>
           <option value="">All Categories</option>
